fix: separated logic from TechnicalPage into a custom hook named useTechnicalTestPage and changed the name of the hook useTechnicalTest to useTechnicalTestList, because it was only usable into TechnicalTestList component
